requirements.txt
streamlit
python-dotenv
groq
langchain-pinecone
langchain-community
langchain-text-splitters
beautifulsoup4
html2text
requests
pydantic
pydantic-settings
